requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-arm64
aiohttp=3.9.5=pypi_0
aiosignal=1.3.1=pypi_0
alembic=1.13.1=pypi_0
annotated-types=0.7.0=pypi_0
anyio=4.3.0=pypi_0
appdirs=1.4.4=pypi_0
appnope=0.1.4=pyhd8ed1ab_0
argilla=1.29.0=pypi_0
asgiref=3.8.1=pypi_0
asttokens=2.4.1=pyhd8ed1ab_0
attrs=23.2.0=pypi_0
backoff=2.2.1=pypi_0
bcrypt=4.1.3=pypi_0
beautifulsoup4=4.12.3=pypi_0
boto3=1.34.131=pypi_0
botocore=1.34.131=pypi_0
brotli=1.1.0=pypi_0
build=1.2.1=pypi_0
bzip2=1.0.8=h80987f9_6
ca-certificates=2024.3.11=hca03da5_0
cachetools=5.3.3=pypi_0
certifi=2024.2.2=pypi_0
cffi=1.16.0=pypi_0
charset-normalizer=3.3.2=pypi_0
chroma-hnswlib=0.7.3=pypi_0
chromadb=0.4.24=pypi_0
click=8.1.7=pypi_0
cohere=5.5.8=pypi_0
coloredlogs=15.0.1=pypi_0
comm=0.2.2=pyhd8ed1ab_0
crewai=0.32.2=pypi_0
crewai-tools=0.3.0=pypi_0
cryptography=42.0.7=pypi_0
dataclasses-json=0.6.6=pypi_0
debugpy=1.6.7=py312h313beb8_0
decorator=5.1.1=pyhd8ed1ab_0
deprecated=1.2.14=pypi_0
deprecation=2.1.0=pypi_0
distro=1.9.0=pypi_0
docstring-parser=0.16=pypi_0
docx2txt=0.8=pypi_0
embedchain=0.1.109=pypi_0
et-xmlfile=1.1.0=pypi_0
exceptiongroup=1.2.0=pyhd8ed1ab_2
executing=2.0.1=pyhd8ed1ab_0
expat=2.6.2=h313beb8_0
fastapi=0.110.3=pypi_0
fastavro=1.9.4=pypi_0
filelock=3.14.0=pypi_0
flatbuffers=24.3.25=pypi_0
frozenlist=1.4.1=pypi_0
fsspec=2024.5.0=pypi_0
gitdb=4.0.11=pypi_0
gitpython=3.1.43=pypi_0
google-api-core=2.19.0=pypi_0
google-auth=2.29.0=pypi_0
google-cloud-aiplatform=1.51.0=pypi_0
google-cloud-bigquery=3.23.0=pypi_0
google-cloud-core=2.4.1=pypi_0
google-cloud-resource-manager=1.12.3=pypi_0
google-cloud-storage=2.16.0=pypi_0
google-crc32c=1.5.0=pypi_0
google-resumable-media=2.7.0=pypi_0
googleapis-common-protos=1.63.0=pypi_0
gptcache=0.1.43=pypi_0
grpc-google-iam-v1=0.13.0=pypi_0
grpcio=1.64.0=pypi_0
grpcio-status=1.62.2=pypi_0
h11=0.14.0=pypi_0
httpcore=1.0.5=pypi_0
httptools=0.6.1=pypi_0
httpx=0.26.0=pypi_0
httpx-sse=0.4.0=pypi_0
huggingface=0.0.1=pypi_0
huggingface-hub=0.23.0=pypi_0
humanfriendly=10.0=pypi_0
idna=3.7=pypi_0
importlib-metadata=7.0.0=pypi_0
importlib-resources=6.4.0=pypi_0
importlib_metadata=7.1.0=hd8ed1ab_0
iniconfig=2.0.0=pypi_0
instructor=1.3.3=pypi_0
ipykernel=6.29.3=pyh3cd1d5f_0
ipython=8.24.0=pyh707e725_0
jedi=0.19.1=pyhd8ed1ab_0
jiter=0.4.2=pypi_0
jmespath=1.0.1=pypi_0
joblib=1.4.2=pypi_0
jsonpatch=1.33=pypi_0
jsonpointer=2.4=pypi_0
jsonref=1.1.0=pypi_0
jupyter_client=8.6.1=pyhd8ed1ab_0
jupyter_core=5.5.0=py312hca03da5_0
kubernetes=29.0.0=pypi_0
lancedb=0.5.7=pypi_0
langchain=0.1.20=pypi_0
langchain-cohere=0.1.5=pypi_0
langchain-community=0.0.38=pypi_0
langchain-core=0.1.52=pypi_0
langchain-openai=0.1.7=pypi_0
langchain-text-splitters=0.0.2=pypi_0
langsmith=0.1.81=pypi_0
libcxx=14.0.6=h848a8c0_0
libffi=3.4.4=hca03da5_1
libsodium=1.0.18=h27ca646_1
lxml=5.2.2=pypi_0
mako=1.3.5=pypi_0
markdown=3.6=pypi_0
markdown-it-py=3.0.0=pypi_0
markupsafe=2.1.5=pypi_0
marshmallow=3.21.2=pypi_0
matplotlib-inline=0.1.7=pyhd8ed1ab_0
mdurl=0.1.2=pypi_0
mmh3=4.1.0=pypi_0
monotonic=1.6=pypi_0
mpmath=1.3.0=pypi_0
multidict=6.0.5=pypi_0
mutagen=1.47.0=pypi_0
mypy-extensions=1.0.0=pypi_0
ncurses=6.4=h313beb8_0
nest-asyncio=1.6.0=pyhd8ed1ab_0
nltk=3.8.1=pypi_0
nodeenv=1.8.0=pypi_0
numpy=1.26.4=pypi_0
oauthlib=3.2.2=pypi_0
onnxruntime=1.18.0=pypi_0
openai=1.30.1=pypi_0
openpyxl=3.1.3=pypi_0
openssl=3.3.0=hfb2fe0b_1
opentelemetry-api=1.24.0=pypi_0
opentelemetry-exporter-otlp-proto-common=1.24.0=pypi_0
opentelemetry-exporter-otlp-proto-grpc=1.24.0=pypi_0
opentelemetry-exporter-otlp-proto-http=1.24.0=pypi_0
opentelemetry-instrumentation=0.45b0=pypi_0
opentelemetry-instrumentation-asgi=0.45b0=pypi_0
opentelemetry-instrumentation-fastapi=0.45b0=pypi_0
opentelemetry-proto=1.24.0=pypi_0
opentelemetry-sdk=1.24.0=pypi_0
opentelemetry-semantic-conventions=0.45b0=pypi_0
opentelemetry-util-http=0.45b0=pypi_0
orjson=3.10.3=pypi_0
outcome=1.3.0.post0=pypi_0
overrides=7.7.0=pypi_0
packaging=23.2=pypi_0
pandas=2.2.2=pypi_0
parameterized=0.9.0=pypi_0
parso=0.8.4=pyhd8ed1ab_0
pexpect=4.9.0=pyhd8ed1ab_0
pickleshare=0.7.5=py_1003
pillow=10.3.0=pypi_0
pip=24.0=py312hca03da5_0
platformdirs=4.2.2=pyhd8ed1ab_0
pluggy=1.5.0=pypi_0
posthog=3.5.0=pypi_0
prompt-toolkit=3.0.42=pyha770c72_0
proto-plus=1.23.0=pypi_0
protobuf=4.25.3=pypi_0
psutil=5.9.0=py312h80987f9_0
ptyprocess=0.7.0=pyhd3deb0d_0
pulsar-client=3.5.0=pypi_0
pure_eval=0.2.2=pyhd8ed1ab_0
py=1.11.0=pypi_0
pyarrow=16.1.0=pypi_0
pyasn1=0.6.0=pypi_0
pyasn1-modules=0.4.0=pypi_0
pycparser=2.22=pypi_0
pycryptodomex=3.20.0=pypi_0
pydantic=2.7.1=pypi_0
pydantic-core=2.18.2=pypi_0
pygithub=1.59.1=pypi_0
pygments=2.18.0=pyhd8ed1ab_0
pyjwt=2.8.0=pypi_0
pylance=0.9.18=pypi_0
pynacl=1.5.0=pypi_0
pypdf=4.2.0=pypi_0
pypika=0.48.9=pypi_0
pyproject-hooks=1.1.0=pypi_0
pyright=1.1.363=pypi_0
pysbd=0.3.4=pypi_0
pysocks=1.7.1=pypi_0
pytest=8.2.1=pypi_0
python=3.12.3=h99e199e_1
python-dateutil=2.9.0.post0=pypi_0
python-docx=1.1.2=pypi_0
python-dotenv=1.0.1=pypi_0
python-magic=0.4.27=pypi_0
python-pptx=0.6.23=pypi_0
pytube=15.0.0=pypi_0
pytz=2024.1=pypi_0
pyyaml=6.0.1=pypi_0
pyzmq=25.1.2=py312h313beb8_0
ratelimiter=1.2.0.post0=pypi_0
readline=8.2=h1a28f6b_0
regex=2023.12.25=pypi_0
requests=2.32.1=pypi_0
requests-oauthlib=2.0.0=pypi_0
retry=0.9.2=pypi_0
rich=13.7.1=pypi_0
rsa=4.9=pypi_0
s3transfer=0.10.1=pypi_0
schema=0.7.7=pypi_0
selenium=4.21.0=pypi_0
semver=3.0.2=pypi_0
setuptools=69.5.1=py312hca03da5_0
shapely=2.0.4=pypi_0
six=1.16.0=pyh6c4a22f_0
smmap=5.0.1=pypi_0
sniffio=1.3.1=pypi_0
sortedcontainers=2.4.0=pypi_0
soupsieve=2.5=pypi_0
sqlalchemy=2.0.30=pypi_0
sqlite=3.45.3=h80987f9_0
stack_data=0.6.2=pyhd8ed1ab_0
starlette=0.37.2=pypi_0
sympy=1.12=pypi_0
tenacity=8.3.0=pypi_0
tiktoken=0.7.0=pypi_0
tk=8.6.14=h6ba3021_0
tmdbv3api=1.9.0=pypi_0
tokenizers=0.19.1=pypi_0
tornado=6.3.3=py312h80987f9_0
tqdm=4.66.4=pypi_0
traitlets=5.14.3=pyhd8ed1ab_0
trio=0.25.1=pypi_0
trio-websocket=0.11.1=pypi_0
typer=0.9.4=pypi_0
types-requests=2.32.0.20240622=pypi_0
typing-inspect=0.9.0=pypi_0
typing_extensions=4.11.0=pyha770c72_0
tzdata=2024.1=pypi_0
unstructured=0.5.1=pypi_0
urllib3=2.2.1=pypi_0
uvicorn=0.29.0=pypi_0
uvloop=0.19.0=pypi_0
watchfiles=0.21.0=pypi_0
wcwidth=0.2.13=pyhd8ed1ab_0
websocket-client=1.8.0=pypi_0
websockets=12.0=pypi_0
wheel=0.43.0=py312hca03da5_0
wrapt=1.14.1=pypi_0
wsproto=1.2.0=pypi_0
xlsxwriter=3.2.0=pypi_0
xz=5.4.6=h80987f9_1
yarl=1.9.4=pypi_0
youtube-transcript-api=0.6.2=pypi_0
yt-dlp=2023.12.30=pypi_0
zeromq=4.3.5=h313beb8_0
zipp=3.18.2=pypi_0
zlib=1.2.13=h18a0788_1
